refactor(backend-cuda): split — extract BackendQuantMarlin + BackendQuantGguf (#8 Phase 3)
Move the largest remaining supertrait body out of `cuda/mod.rs`:

- `cuda/quant.rs` (~1055 lines): `GptqStoreCuda` enum + cfg-gated alias
  + `use_triton_int4`/`use_vllm_moe` env toggles + `MarlinGatherScratch`
  process-global staging buffer (with `pregrow_marlin_gather_scratch`
  free function for graph-capture pre-warm) + `moe_gemm_phase_fused_impl`
  (Stage 12 fused MoE Marlin) + `marlin_gemm_with_perm` perm-aware
  dispatcher + `launch_vllm_marlin` (vendored vLLM marlin path) +
  cfg(not) shim + `impl BackendQuantMarlin for CudaBackend` body +
  `impl BackendQuantGguf for CudaBackend {}` placeholder.

`mod.rs` re-exports `GptqStoreCuda`, `marlin_gemm_with_perm`, and
(under `marlin` feature) `pregrow_marlin_gather_scratch` to preserve
the historical `crate::backend::cuda::*` paths used by
`quant_linear::cuda_marlin` and `tests/marlin_moe_parity_test.rs`.

`default_stream` and `ensure_module` (used by quant.rs) bumped to
`pub(super)`. No behaviour change — pure mechanical extraction.

After Phase 3: mod.rs **4412 → 3388 lines (-23%)**. Cumulative since
Phase 0: 4756 → 3388 (-29%). Phase 4 (next): BackendPagedKv (~666).
